+++ Sources/Internal/Scene3D/Components/SoundComponent.h
@@ -111,7 +111,7 @@ inline SoundEvent * SoundComponent::GetSoundEvent(uint32 index) const
 
 inline uint32 SoundComponent::GetEventsCount() const
 {
-    return events.size();
+    return static_cast<uint32>(events.size());
 }
 
 inline uint32 SoundComponent::GetSoundEventFlags(uint32 index) const
