+++ Sources/Internal/Scene3D/Systems/BaseProcessSystem.cpp
@@ -49,7 +49,7 @@ void BaseProcessSystem::AddComponent( Entity * entity, Component * component )
 
 void BaseProcessSystem::RemoveComponent( Entity * entity, Component * component )
 {
-	uint32 size = components.size();
+	uint32 size = static_cast<uint32>(components.size());
 	for(uint32 i = 0; i < size; ++i)
 	{
 		if(components[i] == component)
